refactor: #386 Service 책임에 따른 분리
- PerformanceService: 기본 CRUD 비즈니스 로직
- PerformanceClientService: FeignClient 비즈니스 로직
