+++ com.taken_seat.performance-service/src/main/java/com/taken_seat/performance_service/performance/presentation/controller/PerformanceController.java
@@ -22,6 +22,7 @@ import com.taken_seat.common_service.dto.ApiResponseData;
 import com.taken_seat.common_service.dto.AuthenticatedUser;
 import com.taken_seat.common_service.dto.response.PerformanceEndTimeDto;
 import com.taken_seat.common_service.dto.response.PerformanceStartTimeDto;
+import com.taken_seat.performance_service.performance.application.service.PerformanceClientService;
 import com.taken_seat.performance_service.performance.application.service.PerformanceService;
 import com.taken_seat.performance_service.performance.presentation.docs.PerformanceControllerDocs;
 import com.taken_seat.performance_service.performance.presentation.dto.request.CreateRequestDto;
@@ -41,6 +42,7 @@ import lombok.RequiredArgsConstructor;
 public class PerformanceController implements PerformanceControllerDocs {
 
 	private final PerformanceService performanceService;
+	private final PerformanceClientService performanceClientService;
 
 	@PostMapping
 	public ResponseEntity<ApiResponseData<CreateResponseDto>> create(
@@ -90,7 +92,8 @@ public class PerformanceController implements PerformanceControllerDocs {
 		@PathVariable("performanceId") UUID performanceId,
 		@PathVariable("performanceScheduleId") UUID performanceScheduleId) {
 
-		PerformanceEndTimeDto response = performanceService.getPerformanceEndTime(performanceId, performanceScheduleId);
+		PerformanceEndTimeDto response = performanceClientService.getPerformanceEndTime(performanceId,
+			performanceScheduleId);
 		return ResponseEntity.ok(ApiResponseData.success(response));
 	}
 
@@ -108,7 +111,7 @@ public class PerformanceController implements PerformanceControllerDocs {
 		@PathVariable("performanceId") UUID performanceId,
 		@PathVariable("performanceScheduleId") UUID performanceScheduleId) {
 
-		PerformanceStartTimeDto response = performanceService.getPerformanceStartTime(performanceId,
+		PerformanceStartTimeDto response = performanceClientService.getPerformanceStartTime(performanceId,
 			performanceScheduleId);
 		return ResponseEntity.ok(ApiResponseData.success(response));
 	}
